Naming support aliyun STS auth.
