+++ client/src/main/java/com/alibaba/nacos/client/auth/ram/injector/NamingResourceInjector.java
@@ -18,6 +18,10 @@ package com.alibaba.nacos.client.auth.ram.injector;
 
 import com.alibaba.nacos.api.common.Constants;
 import com.alibaba.nacos.api.naming.utils.NamingUtils;
+import com.alibaba.nacos.client.auth.ram.identify.IdentifyConstants;
+import com.alibaba.nacos.client.auth.ram.identify.StsConfig;
+import com.alibaba.nacos.client.auth.ram.identify.StsCredential;
+import com.alibaba.nacos.client.auth.ram.identify.StsCredentialHolder;
 import com.alibaba.nacos.plugin.auth.api.LoginIdentityContext;
 import com.alibaba.nacos.client.auth.ram.RamContext;
 import com.alibaba.nacos.plugin.auth.api.RequestResource;
@@ -43,11 +47,20 @@ public class NamingResourceInjector extends AbstractResourceInjector {
     public void doInject(RequestResource resource, RamContext context, LoginIdentityContext result) {
         if (context.validate()) {
             try {
+                String accessKey = context.getAccessKey();
+                String secretKey = context.getSecretKey();
+                // STS 临时凭证鉴权的优先级高于 AK/SK 鉴权
+                if (StsConfig.getInstance().isStsOn()) {
+                    StsCredential stsCredential = StsCredentialHolder.getInstance().getStsCredential();
+                    accessKey = stsCredential.getAccessKeyId();
+                    secretKey = stsCredential.getAccessKeySecret();
+                    result.setParameter(IdentifyConstants.SECURITY_TOKEN_HEADER, stsCredential.getSecurityToken());
+                }
                 String signData = getSignData(getGroupedServiceName(resource));
-                String signature = SignUtil.sign(signData, context.getSecretKey());
+                String signature = SignUtil.sign(signData, secretKey);
                 result.setParameter(SIGNATURE_FILED, signature);
                 result.setParameter(DATA_FILED, signData);
-                result.setParameter(AK_FILED, context.getAccessKey());
+                result.setParameter(AK_FILED, accessKey);
             } catch (Exception e) {
                 NAMING_LOGGER.error("inject ak/sk failed.", e);
             }
